+++ src/main/java/net/sf/jabref/gui/ParserResultWarningDialog.java
@@ -15,14 +15,21 @@
 */
 package net.sf.jabref.gui;
 
+import java.awt.Dimension;
 import java.util.List;
 import java.util.Objects;
 
 import javax.swing.JOptionPane;
+import javax.swing.JScrollPane;
+import javax.swing.JTextArea;
 
 import net.sf.jabref.importer.ParserResult;
 import net.sf.jabref.logic.l10n.Localization;
 
+/**
+ * Class for generating a dialog showing warnings from ParserResult
+ *
+ */
 public class ParserResultWarningDialog {
 
     /**
@@ -31,10 +38,10 @@ public class ParserResultWarningDialog {
      * @param parserResult - ParserResult for the current import/open
      * @param jabRefFrame - the JabRefFrame
      */
-    public static void showParserResultWarningDialog(ParserResult parserResult, JabRefFrame jabRefFrame) {
+    public static void showParserResultWarningDialog(final ParserResult parserResult, final JabRefFrame jabRefFrame) {
         Objects.requireNonNull(parserResult);
         Objects.requireNonNull(jabRefFrame);
-        showParserResultWarningDialog(parserResult, jabRefFrame, Integer.MAX_VALUE, -1);
+        showParserResultWarningDialog(parserResult, jabRefFrame, -1);
     }
 
     /**
@@ -42,11 +49,10 @@ public class ParserResultWarningDialog {
      *
      * @param parserResult - ParserResult for the current import/open
      * @param jabRefFrame - the JabRefFrame
-     * @param maxWarnings - Maximum number of warnings to display
      * @param dataBaseNumber - Database tab number to activate when showing the warning dialog
      */
-    public static void showParserResultWarningDialog(ParserResult parserResult, JabRefFrame jabRefFrame,
-            int maxWarnings, int dataBaseNumber) {
+    public static void showParserResultWarningDialog(final ParserResult parserResult, final JabRefFrame jabRefFrame,
+            final int dataBaseNumber) {
         Objects.requireNonNull(parserResult);
         Objects.requireNonNull(jabRefFrame);
         // Return if no warnings
@@ -60,17 +66,13 @@ public class ParserResultWarningDialog {
         }
 
         // Generate string with warning texts
-        List<String> warnings = parserResult.warnings();
-        StringBuilder dialogContent = new StringBuilder();
-        for (int j = 0; j < Math.min(maxWarnings, warnings.size()); j++) {
-            dialogContent.append(j + 1).append(". ").append(warnings.get(j)).append("\n");
+        final List<String> warnings = parserResult.warnings();
+        final StringBuilder dialogContent = new StringBuilder();
+        int warningCount = 1;
+        for (final String warning : warnings) {
+            dialogContent.append(String.format("%d. %s\n", warningCount++, warning));
         }
-        if (warnings.size() > maxWarnings) {
-            dialogContent.append("... ");
-            dialogContent.append(Localization.lang("%0 warnings", String.valueOf(warnings.size())));
-        } else if (dialogContent.length() > 0) {
-            dialogContent.deleteCharAt(dialogContent.length() - 1);
-        }
+        dialogContent.deleteCharAt(dialogContent.length() - 1);
 
         // Generate dialog title
         String dialogTitle;
@@ -80,15 +82,17 @@ public class ParserResultWarningDialog {
             dialogTitle = Localization.lang("Warnings") + " (" + parserResult.getFile().getName() + ")";
         }
 
-        // Comment from the old code:
-        //
-        // Note to self or to someone else: The following line causes an
-        // ArrayIndexOutOfBoundsException in situations with a large number of
-        // warnings; approx. 5000 for the database I opened when I observed the problem
-        // (duplicate key warnings). I don't think this is a big problem for normal situations,
-        // and it may possibly be a bug in the Swing code.
+        // Create JTextArea with JScrollPane
+        final JTextArea textArea = new JTextArea(dialogContent.toString());
+        final JScrollPane scrollPane = new JScrollPane(textArea) {
+
+            @Override
+            public Dimension getPreferredSize() {
+                return new Dimension(800, Math.min(Math.max(100, warnings.size() * 15), 400)); // Guess a suitable height between 100 and 400
+            }
+        };
 
         // Show dialog
-        JOptionPane.showMessageDialog(jabRefFrame, dialogContent.toString(), dialogTitle, JOptionPane.WARNING_MESSAGE);
+        JOptionPane.showMessageDialog(jabRefFrame, scrollPane, dialogTitle, JOptionPane.WARNING_MESSAGE);
     }
 }